Remove nested-loops test suite.
- Add it to the Prelude test suite.
