+++ streamly.cabal
@@ -583,6 +583,7 @@ test-suite Prelude
     , containers        >= 0.5   && < 0.7
     , transformers      >= 0.4   && < 0.6
     , mtl               >= 2.2   && < 3
+    , random            >= 1.0.0 && < 2
 
 test-suite Prelude.Serial
   import: test-options, prelude-test-options
@@ -815,17 +816,6 @@ test-suite string-test
     , QuickCheck        >= 2.13  && < 2.15
     , hspec             >= 2.0   && < 3
   default-language: Haskell2010
-
-test-suite nested-loops
-  import: test-options
-  type: exitcode-stdio-1.0
-  default-language: Haskell2010
-  main-is: nested-loops.hs
-  hs-source-dirs:  test
-  build-Depends:
-      streamly
-    , base   >= 4.9   && < 5
-    , random >= 1.0.0 && < 2
 
 test-suite parallel-loops
   import: test-options
